Introduction

Inherit.jl is used to inherit fields and interface definitions from a supertype. It supports programming with an object-oriented flavor in Julia, whenever this is more appropriate than developing under traditional Julia patterns.

Fields defined in a supertype are automatically inherited by each subtype, and method declarations are checked for each subtype's implementation. An inheritance hierachy across multiple modules is supported. To accomplish this, macro processing is used to construct native Julia types, which allows the the full range of Julia syntax to be used in most situations.

Quick Start

Use @abstractbase to declare an abstract supertype, and use @implement to inherit from such a type. Standard struct syntax is used.

using Inherit

"abstract base type of Fruity objects"
@abstractbase struct Fruit
	weight::Float64
	"declares an interface which must be implemented"
	function cost(fruit::Fruit, unitprice::Float64) end
end

"
concrete type which represents an apple, inheriting from Fruit
it has two fields: `weight` and `cost`
"
@implement struct Apple <: Fruit 
	coresize::Int
end

"implements supertype's interface declaration `cost` for the type `Apple`"
function cost(apple::Apple, unitprice::Float64)
	apple.weight * unitprice * (apple.coresize < 5 ? 2.0 : 1.0)
end

println(cost(Apple(3.0, 4), 1.0))

# output
6.0

Note that the definition of cost function inside of Fruit is interpreted as an interface declaration; it does not result in a method being defined.

Info

What this declaration means is that when invoking the cost function, passing an object which is a subtype of Fruit (declared with the @implement macro) to the fruit::Fruit parameter must be able to dispatch to some method instance. This is verified when a module is first loaded.

Interaction with modules

The module __init__() function

The @postinit macro

Changing the reporting level

By default, module __init__() writes a summary message at the info log level. You change this by setting ENV["INHERIT_JL_SUMMARY_LEVEL"] to one of ["debug", "info", "warn", "error", "none"]

Limitations

Multiple inheritance

Introduction

Quick Start

Interaction with modules

The module __init__() function

The @postinit macro

Changing the reporting level

Limitations

Multiple inheritance

Inherit.Inherit

Inherit.DB_FLAGS

Inherit.RL

Inherit.createshadowmodule

Inherit.privatize_funcname

Inherit.setglobalreportlevel

Inherit.setreportlevel

Inherit.to_import_expr

Inherit.to_qualified_expr

Inherit.@abstractbase

Inherit.@implement

Inherit.@interface

Inherit.@postinit

Inherit.@test_nothrows

Inherit.Inherit — Module

Inherit.jl let's the user inherit fields and interface definitions from a supertype. There are two macros which provide abstract templates for concrete types: @abstractbase which declares an abstract supertype with inheritable fields and required method definitions, and @interface which declares a container rather than supertype, while providing the same inheritance features. The @implement macro creates the concrete type that implements either or both of these templates. While a concrete type may implement only one @abstractbase, multiple @interface's may be implemented.

We have steered away from using the term "traits" to avoid confusion with extensible Holy traits widely used in Julia libraries. While @interface's can be multiplely inherited, they cannot be added to an existing concrete type from another package.

Limitations

Concrete type must be defined in the same module as the method definitions specific t.

Short form function definitions such as f() = nothing are not supported for method declaration; use the long form function f() end instead. Using short form for method implementation can be problematic as well (e.g. when the function is imported from another module); it's generally safer to use long form.

Just like Julia types, definitions should be given in the order of their dependencies. While out-of-order code can work in some circumstances, we don't test for them. Within a given type, field and method definitions must be unique.

Inherit.jl has no special knowledge about constructors (inner or otherwise). They're treated like normal functions.

The package's macros must be used at the toplevel of a module. @abstractbase relies on world age to advance in order to detect a "real" method was defined (to differentiate the case where the method definition has identical signature as interface specification).

If you cannot return to toplevel (e.g. being wrapped in a @testset macro), a work around is to modify the signature slightly but retain the call paths that you require.

Currently we only handle long form interface definitions such as function f() end.

TODO: multiple levels of interface (not hard) TODO: multiple interfaces (may be hard) TODO: what about parametric types?

A method's signature given only by its positional arguments is unique. If you define a method with the same positional arguments but different keyword arguments from a previously defined method, it will overwrite the previous method. Keyword arguments simply do not particular in method dispatch.

A parametric type signature can be supertype of abstract type signature Tuple{typeof(f), Real} <: Tuple{typeof(f), T} where T<:Number

typeof(f).name.mt grows for each evaluation of method definition, even if it overwrites a previous definition. It is not the same as methods(f)

Inherit.DB_FLAGS — Constant

This gives us a list of all modules that are actively using Inherit.jl, if we ever need it

Inherit.RL — Type

ThrowError: Checks for interface requirements in the module __init__() function, and throws an exception if requirements are not met. Creates __init__ upon first use of @abstractbase or @implement.

ShowMessage: Same as above, but an error level log message is shown rather than throwing exception.

DisableInit: Will not create a module __init__() function; disables the @postinit macro. Can not be set if __init__ has already been created

Inherit.createshadowmodule — Method

Creates a new module that contains (i.e. imports) only the properties of basemodule which are Types and Modules (i.e. excluding any functions). You can evaluate method declarations in this module to get the signature, without modifying basemodule itself

Inherit.privatize_funcname — Method

prepend __ to function name

Inherit.setglobalreportlevel — Method

This sets the default reporting level of modules; it does not change the setting for modules that already have a ModuleEntry.

Inherit.setreportlevel — Method

Takes precedence over global report level.

Inherit.to_import_expr — Method

Makes a valid import expression like import modpath1.modpath2.modpath3: item

:(import $modpath : $item) won't work even when modpath evaluates to modpath1.modpath2.modpath3

Inherit.to_qualified_expr — Method

converts syntax like a.b.c.d to AST

Inherit.@abstractbase — Macro

TODO: evaluate in a temporary module TODO: support mutable types

Inherit.@implement — Macro

Method declarations may come from a foreign module, in which case, method implementations must belong to functions in that foreign module. If there's no name clash, the foreign modules's function is automatically imported into the implementing module (i.e. your current module). If there is a name clash, you must qualify the method implementation with the foreign module's name.

Inherit.@interface — Macro

An @abstractbase follows Julia's type hierarchy; a concrete type may only implement one abstractbase. A @interface is similar in some ways to Holy traits; a type may implement multiple interfaces in addition to its abstractbase. A interface can span type hierarchies, but it may only be used to inherit fields and function definition requirements. It cannot be used as a container element or object type (while carrying the behavior of interfaces).

Can recreate the struct parameterized by the interface, this allows dispatch only on type, or on both type and interface. Basically, store the list of interfaces in the type parameters, and create default constructors that don't require the interface parameters.

Inherit.@postinit — Macro

Executed after Inherit.jl verfies interfaces. You may have any number of @postinit blocks; they will execute in the sequence order in which they're defined.

Inherit.@test_nothrows — Macro

opposite of @test_throws